chapter_002/src/test/java/ru/job4j/pseudo/SquareTest.java
package ru.job4j.pseudo;

import org.junit.Test;
import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

/**
 * @author Aleksey Vdovenko (a.a.vd0venk0@gmail.com)
 * @version $Id$
 * @since 0.1
 */
public class SquareTest {
    /**
     * Тест для квадрата.
     */
    @Test
    public void whenDrawSquare() {
        Square square = new Square();
        assertThat(
                square.draw(),
                is(
                        new StringBuilder()
                        .append("*****\n")
                        .append("*   *\n")
                        .append("*   *\n")
                        .append("*   *\n")
                        .append("*****\n")
                        .toString()
                )
        );
    }
}
